requirements.txt
python-Levenshtein==0.27.1
transformers==4.50.1
torch==2.7.1+cu118
torchvision==0.22.1+cu118
torchaudio==2.7.1+cu118
pdfplumber==0.9.0
python-docx==0.8.11
django==4.2.3
djangorestframework==3.14.0
psycopg2-binary==2.9.6
django-cors-headers==4.2.0 